Implementando DTO's

+++ backend/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers;
 
@@ -21,7 +21,8 @@ public class EventosController : ControllerBase
         try
         {
             var eventos = await _eventoService.GetAllEventosAsync(true);
-            if (eventos == null) return NotFound("Nenhum evento encontrado.");
+            if (eventos == null) return NoContent();
+
             return Ok(eventos);
         }
         catch (Exception ex)
@@ -39,7 +40,7 @@ public class EventosController : ControllerBase
         try
         {
             var evento = await _eventoService.GetEventoByIdAsync(id, true);
-            if (evento == null) return NotFound("Nenhum evento encontrado.");
+            if (evento == null) return NoContent();
             return Ok(evento);
         }
         catch (Exception ex)
@@ -57,7 +58,7 @@ public class EventosController : ControllerBase
         try
         {
             var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-            if (eventos == null) return NotFound("Nenhum evento encontrado.");
+            if (eventos == null) return NoContent();
             return Ok(eventos);
         }
         catch (Exception ex)
@@ -70,7 +71,7 @@ public class EventosController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> Post(Evento model)
+    public async Task<IActionResult> Post(EventoDto model)
     {
         try
         {
@@ -88,7 +89,7 @@ public class EventosController : ControllerBase
     }
 
     [HttpPut("{id}")]
-    public async Task<IActionResult> Put(int id, Evento model)
+    public async Task<IActionResult> Put(int id, EventoDto model)
     {
         try
         {
@@ -110,8 +111,10 @@ public class EventosController : ControllerBase
     {
         try
         {
+            var evento = await _eventoService.GetEventoByIdAsync(id);
+            if (evento == null) return NoContent();
             return await _eventoService.DeleteEvento(id) ?
-                Ok("Evento deletado.") : BadRequest("Evento não deletado");
+                Ok("Evento deletado.") : throw new Exception("Erro ao deletar evento.");
         }
         catch (Exception ex)
         {
